Address all PR review comments
- #1: docstring "on" → "for" Pydantic AI
- #2: re-export CodeMode from top-level package
- #3: single-backtick Markdown in all docstrings
- #4/5: replace double FilteredToolset with internal filtering via
  tool_selector + matches_tool_selector in CodeModeToolset
- #6: remove empty environments/__init__.py
- #7: rename toolsets/code_mode → toolsets/_code_mode (private)
- #9: inline code_mode/__init__.py
- #10: move "async" notice to _FUNCTIONS_HEADER
- #11: make max_retries configurable on CodeMode + CodeModeToolset
- #12: _RunCodeTool(ToolsetTool) subclass caches callable_defs,
  sanitized_to_original, wrapped_tools — no duplicate get_tools call
- #13: extracted repeated .get() into var
- #15: separate warnings for deferred execution vs deferred loading,
  added TODO for search_tool exclusion
- #16/17: collision check catches native-overwrites-sanitized case
- #18/19: use conflicting_type_names param on render(), removed
  _render_tool_name ContextVar and manual type rendering helpers
- #20: dispatch sandboxed tool calls through ToolManager.handle_call()
  via ctx.tool_manager (with fallback for tests/TemporalRunContext)
- #22: move logfire[httpx] from dependencies to dev dependency

Depends on pydantic-ai commit 641530a7 which adds tool_manager to
RunContext.
