--- pydantic_harness/environments/__init__.py
@@ -1,1 +1,0 @@
-"""Execution environments used by harness capabilities (e.g. sandboxes, runtimes)."""
