--- pydantic_harness/toolsets/code_mode/__init__.py
@@ -1,5 +1,0 @@
-"""Code mode toolset: routes LLM-generated Python through a Monty sandbox."""
-
-from .run_code import CodeModeToolset
-
-__all__ = ['CodeModeToolset']
